Fix invisible ImporterMeshInstance3D instances after import
The vrm_secondary.skeleton setter was adding an internal gizmo node during import.
This happened response to the automatic property modification when renaming Skeleton3D to GeneralSkeleton
Godot has a bug in which Node.replace_by fails to preserve internal_mode and calculates the incorrect index.
This combination caused the for loop in ResourceImporterScene to skip over some importer meshes.